自定义神化词缀，无线ae删除按钮添加防误触 https://github.com/GregTech-Odyssey/GregTech-Odyssey/issues/1065

+++ src/main/java/com/gtocore/common/machine/multiblock/electric/adventure/SlaughterhouseMachine.java
@@ -244,7 +244,7 @@ public final class SlaughterhouseMachine extends StorageMultiblockMachine implem
                 }
                 if (!(entity instanceof Mob mob)) continue;
                 if (CommonProxy.isBoss(entity)) continue;
-                xp += mob.getExperienceReward() * multiplier;
+                xp += (long) mob.getExperienceReward() * multiplier;
                 getAllDeathLoot(player, serverLevel, mob, source, lootParams, itemStacks, multiplier);
             }
             if (xp > 0) outputFluid(EIOFluids.XP_JUICE.getSource(), xp);
